+++ javascripts/connexion.js
@@ -8,7 +8,6 @@ function performLogin() {
         url: 'functions/login', // Remplacez ceci par le chemin de votre script côté serveur
         data: { identifiant: identifiant, password: password },
         success: function (response) {
-            console.log(response);
             // Afficher le résultat de la connexion
             if (response == "succes") {
                 showPopup("good", "YOUHOU !!", "Vous êtes connecté, Redirection dans 4 Secondes, pas 3, pas 5, 4 ! ... Ouais je suis précis.");