+++ website/agent/dash_agent.php
@@ -15,7 +15,7 @@ if (!isset($uuid) || $uuid == null || $uuid == "uuid") {
     $req->execute($v);
     $result = $req->rowCount();
     if ($result == 0) {
-        return null;
+        $agent = array('agent-name' => 'NOT FOUND');
     } else {
         foreach ($req as $agent) {
         };
@@ -23,104 +23,121 @@ if (!isset($uuid) || $uuid == null || $uuid == "uuid") {
 }
 ?>
 
-<script>
-    document.title = "Agent - <?= $agent['agent_name'] ?>";
-</script>
 
 
 
-<div class="content-dashboard-agent">
-    <section>
-        <h2>Informations Agent</h2>
-        <div class="agent-data">
-            <p class="label">Nom de l'agent:</p>
-            <p class="data-agent"><?= $agent['agent_name'] ?></p>
-        </div>
-        <div class="agent-data">
-            <p class="label">UUID:</p>
-            <p id="agent-uuid" class="data-agent"><?= $agent['uuid_agent'] ?></p>
-        </div>
-        <div class="agent-data">
-            <p class="label">Etat de l'agent:</p>
-            <p id="agent-statut" class="data-agent">Non disponible</p>
-        </div>
-    </section>
-    <?php if ($agent['agent_etat'] !== 0) : ?>
-        <section id="config-sys">
-            <h2>Configuration Système</h2>
+
+<div id="Block-Section-Data" class="content-dashboard-agent">
+
+    <?php if ($result == 0) : ?>
+        <section>
+            <h2>Informations Agent</h2>
+            <p>L'agent n'existe pas. UUID Not found</p>
+        </section>
+    <?php else : ?>
+        <script>
+            document.title = "Agent - <?= $agent['agent_name'] ?>";
+        </script>
+        <section>
+            <h2>Informations Agent</h2>
             <div class="agent-data">
-                <p class="label">Model Processor:</p>
-                <p id="data-agent-ModelProcessor" class="data-agent">Chargement des données</p>
+                <p class="label">Nom de l'agent:</p>
+                <p class="data-agent"><?= $agent['agent_name'] ?></p>
             </div>
             <div class="agent-data">
-                <p class="label">Architecture:</p>
-                <p id="data-agent-architecture" class="data-agent">Chargement des données</p>
+                <p class="label">UUID:</p>
+                <p id="agent-uuid" class="data-agent"><?= $agent['uuid_agent'] ?></p>
             </div>
             <div class="agent-data">
-                <p class="label">Coeurs:</p>
-                <p id="data-agent-Cores" class="data-agent">Chargement des données</p>
-            </div>
-            <div class="agent-data">
-                <p class="label">Utiliation:</p>
-                <p id="data-agent-Used" class="data-agent">Chargement des données</p>
-            </div>
-            <div class="agent-data">
-                <p class="label">Dernière mise à jour:</p>
-                <p id="data-agent-processorUpdate" class="data-agent">Chargement des données</p>
-            </div>
-            <div class="agent-data">
-                <p class="label">Alerte:</p>
-                <p id="config-sys-alerte" class="data-agent alert">RAS</p>
+                <p class="label">Etat de l'agent:</p>
+                <p id="agent-statut" class="data-agent">Non disponible</p>
             </div>
         </section>
+        <?php if ($agent['agent_etat'] !== 0) : ?>
+            <section id="config-sys" <?php if($agent['ProcessorModule_autostart'] != 1){ echo 'style="background-color: grey"';} ?>>
+                <h2>Configuration Système</h2>
+                <div class="agent-data">
+                    <p class="label">Model Processor:</p>
+                    <p id="data-agent-ModelProcessor" class="data-agent">Chargement des données</p>
+                </div>
+                <div class="agent-data">
+                    <p class="label">Architecture:</p>
+                    <p id="data-agent-architecture" class="data-agent">Chargement des données</p>
+                </div>
+                <div class="agent-data">
+                    <p class="label">Coeurs:</p>
+                    <p id="data-agent-Cores" class="data-agent">Chargement des données</p>
+                </div>
+                <div class="agent-data">
+                    <p class="label">Utiliation:</p>
+                    <p id="data-agent-Used" class="data-agent">Chargement des données</p>
+                </div>
+                <div class="agent-data">
+                    <p class="label">Dernière mise à jour:</p>
+                    <p id="data-agent-processorUpdate" class="data-agent">Chargement des données</p>
+                </div>
+                <div class="agent-data">
+                    <p class="label">Alerte:</p>
+                    <p id="config-sys-alerte" class="data-agent alert">RAS</p>
+                </div>
+                <div class="agent-data">
 
-        <section id="memory-sys">
-            <h2>Mémoire système</h2>
-            <div class="agent-data">
-                <p class="label">Mémoire restante:</p>
-                <p id="data-agent-Memfree" class="data-agent">Chargement des données</p>
-            </div>
-            <div class="agent-data">
-                <p class="label">Mémoire total:</p>
-                <p id="data-agent-MemTotal" class="data-agent">Chargement des données</p>
-            </div>
-            <div class="agent-data">
-                <p class="label">Dernière mise à jour:</p>
-                <p id="data-agent-memoryUpdate" class="data-agent">Chargement des données</p>
-            </div>
-            <div class="agent-data">
-                <p class="label">Alerte:</p>
-                <p id="memory-sys-alerte" class="data-agent alert">RAS</p>
-            </div>
-        </section>
+                    <button <?php if($agent['ProcessorModule_autostart'] == 1){ echo 'class="delete"';} ?>  id="ProcessorModule_autostart" data-autostart="<?= $agent['ProcessorModule_autostart'] ?>" onclick="AutoStart('ProcessorModule_autostart')"><?php if($agent['ProcessorModule_autostart'] == 1){ echo 'Désactivé';}else{echo 'Activé';} ?> AutoStart</button>
 
-        <section id="disk-sys">
-            <h2>Disque système</h2>
-            <div class="agent-data">
-                <p class="label">Disques :</p>
-                <p id="data-agent-Memfree" class="data-agent">Non disponible</p>
-            </div>
-            <div class="agent-data">
-                <p class="label">Dernière mise à jour:</p>
-                <p id="data-agent-memoryUpdate" class="data-agent">Non disponible</p>
-            </div>
-        </section>
+                    <button <?php if($agent['ProcessorModule_autorestart'] == 1){ echo 'class="delete"';} ?>  id="ProcessorModule_autorestart" data-autorestart="<?= $agent['ProcessorModule_autorestart'] ?>" style="margin-left: 15px;" onclick="AutoRestart('ProcessorModule_autorestart')"><?php if($agent['ProcessorModule_autorestart'] == 1){ echo 'Désactivé';}else{echo 'Activé';} ?> AutoRestart</button>
 
-        <section id="plex-sys">
-            <h2>Services</h2>
-            <div class="agent-data">
-                <p class="label">Plex :</p>
-                <p id="data-agent-plex" class="data-agent">Chargement des données</p>
-            </div>
-            <div class="agent-data">
-                <p class="label">Dernière mise à jour:</p>
-                <p id="data-agent-servicesUpdate" class="data-agent">Chargement des données</p>
-            </div>
-            <div class="agent-data">
-                <p class="label">Alerte:</p>
-                <p id="plex-sys-alerte" class="data-agent alert">RAS</p>
-            </div>
-        </section>
+                </div>
+            </section>
+
+            <section id="memory-sys" <?php if($agent['MemoryModule_autostart'] != 1){ echo 'style="background-color: grey"';} ?>>
+                <h2>Mémoire système</h2>
+                <div class="agent-data">
+                    <p class="label">Mémoire restante:</p>
+                    <p id="data-agent-Memfree" class="data-agent">Chargement des données</p>
+                </div>
+                <div class="agent-data">
+                    <p class="label">Mémoire total:</p>
+                    <p id="data-agent-MemTotal" class="data-agent">Chargement des données</p>
+                </div>
+                <div class="agent-data">
+                    <p class="label">Dernière mise à jour:</p>
+                    <p id="data-agent-memoryUpdate" class="data-agent">Chargement des données</p>
+                </div>
+                <div class="agent-data">
+                    <p class="label">Alerte:</p>
+                    <p id="memory-sys-alerte" class="data-agent alert">RAS</p>
+                </div>
+                <div class="agent-data">
+
+<button <?php if($agent['MemoryModule_autostart'] == 1){ echo 'class="delete"';} ?>  id="MemoryModule_autostart" data-autostart="<?= $agent['MemoryModule_autostart'] ?>" onclick="AutoStart('MemoryModule_autostart')"><?php if($agent['MemoryModule_autostart'] == 1){ echo 'Désactivé';}else{echo 'Activé';} ?> AutoStart</button>
+
+<button <?php if($agent['MemoryModule_autorestart'] == 1){ echo 'class="delete"';} ?>  id="MemoryModule_autorestart" data-autorestart="<?= $agent['MemoryModule_autorestart'] ?>" style="margin-left: 15px;" onclick="AutoRestart('MemoryModule_autorestart')"><?php if($agent['MemoryModule_autorestart'] == 1){ echo 'Désactivé';}else{echo 'Activé';} ?> AutoRestart</button>
+
+</div>
+            </section>
+
+            <section id="plex-sys" <?php if($agent['PlexProcessCheck_autostart'] != 1){ echo 'style="background-color: grey"';} ?>>
+                <h2>Services</h2>
+                <div class="agent-data">
+                    <p class="label">Plex :</p>
+                    <p id="data-agent-plex" class="data-agent">Chargement des données</p>
+                </div>
+                <div class="agent-data">
+                    <p class="label">Dernière mise à jour:</p>
+                    <p id="data-agent-servicesUpdate" class="data-agent">Chargement des données</p>
+                </div>
+                <div class="agent-data">
+                    <p class="label">Alerte:</p>
+                    <p id="plex-sys-alerte" class="data-agent alert">RAS</p>
+                </div>
+                <div class="agent-data">
+
+                    <button <?php if($agent['PlexProcessCheck_autostart'] == 1){ echo 'class="delete"';} ?>  id="PlexProcessCheck_autostart" data-autostart="<?= $agent['PlexProcessCheck_autostart'] ?>" onclick="AutoStart('PlexProcessCheck_autostart')"><?php if($agent['PlexProcessCheck_autostart'] == 1){ echo 'Désactivé';}else{echo 'Activé';} ?> AutoStart</button>
+
+                    <button <?php if($agent['PlexProcessCheck_autorestart'] == 1){ echo 'class="delete"';} ?>  id="PlexProcessCheck_autorestart" data-autorestart="<?= $agent['PlexProcessCheck_autorestart'] ?>" style="margin-left: 15px;" onclick="AutoRestart('PlexProcessCheck_autorestart')"><?php if($agent['PlexProcessCheck_autorestart'] == 1){ echo 'Désactivé';}else{echo 'Activé';} ?> AutoRestart</button>
+
+                    </div>
+            </section>
 </div>
 
 <section id="block-disk" class="content-dashboard-agent-disk">
@@ -131,5 +148,6 @@ if (!isset($uuid) || $uuid == null || $uuid == "uuid") {
     <section class="info-popup">
         <p>L'agent n'est pas encore associé. Si vous l'avez démarré et renseigné sont UUID, alors vous n'avez simplement qu'a patientez le temps de la synchro.</p>
     </section>
-<?php endif; ?>
+<?php endif;
+    endif; ?>
 <script src="<?= SITE_HTTP . SITE_URL ?>/javascripts/agent/dash-agent.js?0"></script>