+++ website/luma/statut.php
@@ -1,0 +1,24 @@
+<?php
+require('./base/nexus_base.php');
+?>
+<link rel="stylesheet" href="../css/statut.css?1">
+<script>
+    document.title = "Status";
+</script>
+
+<section>
+    <h5>API Nino (PROD)</h5>
+    <p><span id="api-nino-indiq-stats" class=""></span><span id="api-nino-txt-stats">Calcule en cours</span></p>
+</section>
+
+<section>
+    <h5>Plex</h5>
+    <p><span id="plex-indiq-stats" class=""></span><span id="plex-txt-stats">Calcule en cours</span></p>
+</section>
+
+<section>
+    <h5>Serveur Minecraft</h5>
+    <p><span id="indiq-stats" class=""></span><span id="txt-stats">Comming soon</span></p>
+</section>
+
+<script src="../javascripts/home/statut.js"></script>